Extract shell event store
