+++ clients/apple/AlanNative/ShellControlPlane.swift
@@ -49,15 +49,13 @@ final class AlanShellControlPlane {
     private let commandsURL: URL
     private let resultsURL: URL
     private let stateFileURL: URL
-    private let eventsFileURL: URL
     private let commandHandler: (AlanShellControlCommand) -> AlanShellControlResponse
     private let stateAdoptionHandler: @MainActor (ShellStateSnapshot) -> Void
-    private let diagnosticHandler: @MainActor (String) -> Void
+    private let diagnostics: AlanShellDiagnostics
     private let socketServer: AlanShellSocketServer
+    private let eventStore: AlanShellEventStore
     private var filePoller: AlanShellControlFilePoller?
     private var trackedPaneIDs: Set<String> = []
-    private var events: [AlanShellEventEnvelope] = []
-    private var nextEventOrdinal = 1
 
     init(
         windowID: String,
@@ -78,10 +76,16 @@ final class AlanShellControlPlane {
         self.commandsURL = rootURL.appendingPathComponent("commands", isDirectory: true)
         self.resultsURL = rootURL.appendingPathComponent("results", isDirectory: true)
         self.stateFileURL = rootURL.appendingPathComponent("state.json")
-        self.eventsFileURL = rootURL.appendingPathComponent("events.jsonl")
         self.commandHandler = commandHandler
         self.stateAdoptionHandler = stateAdoptionHandler
-        self.diagnosticHandler = diagnosticHandler
+        self.diagnostics = AlanShellDiagnostics(handler: diagnosticHandler)
+        self.eventStore = AlanShellEventStore(
+            windowID: windowID,
+            fileManager: fileManager,
+            eventsFileURL: rootURL.appendingPathComponent("events.jsonl"),
+            encoder: encoder,
+            diagnosticHandler: diagnostics.record
+        )
         self.socketServer = AlanShellSocketServer(
             socketURL: self.socketURL,
             commandHandler: commandHandler,
@@ -105,7 +109,7 @@ final class AlanShellControlPlane {
                 return self.responseForPolledCommand(command)
             },
             bindingProjectionHandler: bindingProjectionHandler,
-            diagnosticHandler: diagnosticHandler
+            diagnosticHandler: diagnostics.record
         )
 
         ensureDirectories()
@@ -142,12 +146,12 @@ final class AlanShellControlPlane {
         let mergeResult = socketServer.mergePublishedState(state)
         let mergedState = mergeResult.merged
         synchronizePaneSupportDirectories(for: mergedState)
-        recordEvents(from: mergeResult.previous, to: mergedState)
+        eventStore.recordChanges(from: mergeResult.previous, to: mergedState)
         do {
             let data = try encoder.encode(mergedState)
             try data.write(to: stateFileURL, options: .atomic)
         } catch {
-            recordDiagnostic("Failed to persist shell state: \(error.localizedDescription)")
+            diagnostics.record("Failed to persist shell state: \(error.localizedDescription)")
         }
     }
 
@@ -156,18 +160,14 @@ final class AlanShellControlPlane {
             do {
                 try fileManager.createDirectory(at: url, withIntermediateDirectories: true)
             } catch {
-                recordDiagnostic("Failed to create shell control directory \(url.path): \(error.localizedDescription)")
+                diagnostics.record("Failed to create shell control directory \(url.path): \(error.localizedDescription)")
             }
         }
     }
 
-    private func recordDiagnostic(_ message: String) {
-        diagnosticHandler(message)
-    }
-
     func specialCommandResponse(for command: AlanShellControlCommand) -> AlanShellControlResponse? {
         guard command.command == .eventsRead else { return nil }
-        let rows = readEvents(afterEventID: command.afterEventID, limit: command.limit)
+        let rows = eventStore.read(afterEventID: command.afterEventID, limit: command.limit)
         return AlanShellControlResponse(
             requestID: command.requestID,
             contractVersion: "0.1",
@@ -187,7 +187,7 @@ final class AlanShellControlPlane {
             acceptedBytes: nil,
             deliveryCode: nil,
             runtimePhase: nil,
-            latestEventID: events.last?.eventID,
+            latestEventID: eventStore.latestEventID,
             errorCode: nil,
             errorMessage: nil
         )
@@ -208,27 +208,12 @@ final class AlanShellControlPlane {
         paneID: String,
         delivery: TerminalRuntimeDeliveryResult
     ) {
-        var payload: [String: AlanShellJSONValue] = [
-            "request_id": .string(requestID),
-            "delivery_code": .string(delivery.code.rawValue),
-            "accepted_bytes": .number(Double(delivery.acceptedBytes))
-        ]
-        if let errorCode = delivery.errorCode {
-            payload["error_code"] = .string(errorCode)
-        }
-        if let errorMessage = delivery.errorMessage {
-            payload["error_message"] = .string(errorMessage)
-        }
-        if let runtimePhase = delivery.runtimePhase {
-            payload["runtime_phase"] = .string(runtimePhase)
-        }
-
-        appendEvent(
-            type: "pane.text_delivery",
+        eventStore.recordTextDelivery(
+            requestID: requestID,
             spaceID: spaceID,
             tabID: tabID,
             paneID: paneID,
-            payload: payload
+            delivery: delivery
         )
     }
 
@@ -247,7 +232,7 @@ final class AlanShellControlPlane {
             do {
                 try fileManager.createDirectory(at: paneURL, withIntermediateDirectories: true)
             } catch {
-                recordDiagnostic("Failed to create pane support directory \(paneURL.path): \(error.localizedDescription)")
+                diagnostics.record("Failed to create pane support directory \(paneURL.path): \(error.localizedDescription)")
             }
         }
 
@@ -260,228 +245,9 @@ final class AlanShellControlPlane {
             do {
                 try fileManager.removeItem(at: paneURL)
             } catch {
-                recordDiagnostic("Failed to remove stale pane support directory \(paneURL.path): \(error.localizedDescription)")
+                diagnostics.record("Failed to remove stale pane support directory \(paneURL.path): \(error.localizedDescription)")
             }
         }
-    }
-
-    private func recordEvents(from previousState: ShellStateSnapshot?, to currentState: ShellStateSnapshot) {
-        guard let previousState else { return }
-
-        let previousPanesByID = Dictionary(uniqueKeysWithValues: previousState.panes.map { ($0.paneID, $0) })
-        let currentPanesByID = Dictionary(uniqueKeysWithValues: currentState.panes.map { ($0.paneID, $0) })
-
-        if previousState.focusedPaneID != currentState.focusedPaneID {
-            appendEvent(
-                type: "focus.changed",
-                spaceID: currentState.focusedSpaceID,
-                tabID: currentState.focusedTabID,
-                paneID: currentState.focusedPaneID,
-                payload: [
-                    "previous_pane_id": .string(previousState.focusedPaneID ?? ""),
-                    "current_pane_id": .string(currentState.focusedPaneID ?? "")
-                ]
-            )
-        }
-
-        let previousTabs = Set(previousState.spaces.flatMap(\.tabs).map(\.tabID))
-        let currentTabs = Set(currentState.spaces.flatMap(\.tabs).map(\.tabID))
-        for createdTabID in currentTabs.subtracting(previousTabs).sorted() {
-            if let tab = currentState.tab(tabID: createdTabID),
-               let paneID = tab.paneTree.paneIDs.first,
-               let pane = currentPanesByID[paneID] {
-                appendEvent(
-                    type: "tab.created",
-                    spaceID: pane.spaceID,
-                    tabID: tab.tabID,
-                    paneID: paneID,
-                    payload: [
-                        "tab_id": .string(tab.tabID),
-                        "kind": .string(tab.kind.rawValue)
-                    ]
-                )
-            }
-        }
-        for closedTabID in previousTabs.subtracting(currentTabs).sorted() {
-            let pane = previousState.panes.first { $0.tabID == closedTabID }
-            appendEvent(
-                type: "tab.closed",
-                spaceID: pane?.spaceID,
-                tabID: closedTabID,
-                paneID: pane?.paneID,
-                payload: ["tab_id": .string(closedTabID)]
-            )
-        }
-
-        let allPaneIDs = Set(previousPanesByID.keys).union(currentPanesByID.keys)
-        for paneID in allPaneIDs.sorted() {
-            let previousPane = previousPanesByID[paneID]
-            let currentPane = currentPanesByID[paneID]
-
-            if let previousPane, let currentPane {
-                if previousPane.tabID != currentPane.tabID || previousPane.spaceID != currentPane.spaceID {
-                    appendEvent(
-                        type: "pane.moved",
-                        spaceID: currentPane.spaceID,
-                        tabID: currentPane.tabID,
-                        paneID: currentPane.paneID,
-                        payload: [
-                            "previous_space_id": .string(previousPane.spaceID),
-                            "current_space_id": .string(currentPane.spaceID),
-                            "previous_tab_id": .string(previousPane.tabID),
-                            "current_tab_id": .string(currentPane.tabID)
-                        ]
-                    )
-                }
-
-                var changedFields: [String] = []
-                if previousPane.cwd != currentPane.cwd {
-                    changedFields.append("cwd")
-                }
-                if previousPane.viewport?.title != currentPane.viewport?.title {
-                    changedFields.append("viewport.title")
-                }
-                if previousPane.viewport?.summary != currentPane.viewport?.summary {
-                    changedFields.append("viewport.summary")
-                }
-                if previousPane.context?.gitBranch != currentPane.context?.gitBranch {
-                    changedFields.append("context.git_branch")
-                }
-                if previousPane.context?.lastCommandExitCode != currentPane.context?.lastCommandExitCode {
-                    changedFields.append("context.last_command_exit_code")
-                }
-                if previousPane.context?.rendererPhase != currentPane.context?.rendererPhase {
-                    changedFields.append("context.renderer_phase")
-                }
-                if previousPane.context?.displayName != currentPane.context?.displayName {
-                    changedFields.append("context.display_name")
-                }
-                if previousPane.context?.displayID != currentPane.context?.displayID {
-                    changedFields.append("context.display_id")
-                }
-                if previousPane.context?.windowTitle != currentPane.context?.windowTitle {
-                    changedFields.append("context.window_title")
-                }
-                if previousPane.context?.socketPath != currentPane.context?.socketPath {
-                    changedFields.append("context.socket_path")
-                }
-                if previousPane.context?.launchCommand != currentPane.context?.launchCommand {
-                    changedFields.append("context.launch_command")
-                }
-                if !changedFields.isEmpty {
-                    appendEvent(
-                        type: "pane.metadata_changed",
-                        spaceID: currentPane.spaceID,
-                        tabID: currentPane.tabID,
-                        paneID: currentPane.paneID,
-                        payload: [
-                            "changed_fields": .array(changedFields.map(AlanShellJSONValue.string))
-                        ]
-                    )
-                }
-
-                if previousPane.attention != currentPane.attention {
-                    appendEvent(
-                        type: "attention.changed",
-                        spaceID: currentPane.spaceID,
-                        tabID: currentPane.tabID,
-                        paneID: currentPane.paneID,
-                        payload: [
-                            "previous": .string(previousPane.attention.rawValue),
-                            "current": .string(currentPane.attention.rawValue)
-                        ]
-                    )
-                }
-
-                if previousPane.alanBinding != currentPane.alanBinding {
-                    appendEvent(
-                        type: "AlanBinding.changed",
-                        spaceID: currentPane.spaceID,
-                        tabID: currentPane.tabID,
-                        paneID: currentPane.paneID,
-                        payload: [
-                            "session_id": .string(currentPane.alanBinding?.sessionID ?? ""),
-                            "run_status": .string(currentPane.alanBinding?.runStatus ?? ""),
-                            "pending_yield": .bool(currentPane.alanBinding?.pendingYield ?? false)
-                        ]
-                    )
-                }
-            } else if let currentPane {
-                appendEvent(
-                    type: "pane.created",
-                    spaceID: currentPane.spaceID,
-                    tabID: currentPane.tabID,
-                    paneID: currentPane.paneID,
-                    payload: [
-                        "pane_id": .string(currentPane.paneID),
-                        "tab_id": .string(currentPane.tabID)
-                    ]
-                )
-            } else if let previousPane {
-                appendEvent(
-                    type: "pane.closed",
-                    spaceID: previousPane.spaceID,
-                    tabID: previousPane.tabID,
-                    paneID: previousPane.paneID,
-                    payload: [
-                        "pane_id": .string(previousPane.paneID)
-                    ]
-                )
-            }
-        }
-    }
-
-    private func appendEvent(
-        type: String,
-        spaceID: String?,
-        tabID: String?,
-        paneID: String?,
-        payload: [String: AlanShellJSONValue]
-    ) {
-        let event = AlanShellEventEnvelope(
-            eventID: "ev_\(nextEventOrdinal)",
-            type: type,
-            timestamp: ISO8601DateFormatter().string(from: .now),
-            windowID: windowID,
-            spaceID: spaceID,
-            tabID: tabID,
-            paneID: paneID,
-            payload: payload
-        )
-        nextEventOrdinal += 1
-        events.append(event)
-        if events.count > 200 {
-            events.removeFirst(events.count - 200)
-        }
-        if let data = try? encoder.encode(event),
-           let line = String(data: data, encoding: .utf8) {
-            do {
-                if fileManager.fileExists(atPath: eventsFileURL.path) {
-                    let handle = try FileHandle(forWritingTo: eventsFileURL)
-                    defer { try? handle.close() }
-                    _ = try handle.seekToEnd()
-                    try handle.write(contentsOf: Data("\(line)\n".utf8))
-                } else {
-                    try Data("\(line)\n".utf8).write(to: eventsFileURL, options: .atomic)
-                }
-            } catch {
-                recordDiagnostic("Failed to persist shell event log: \(error.localizedDescription)")
-            }
-        }
-    }
-
-    private func readEvents(afterEventID: String?, limit: Int?) -> [AlanShellEventEnvelope] {
-        let startIndex: Int
-        if let afterEventID,
-           let index = events.firstIndex(where: { $0.eventID == afterEventID }) {
-            startIndex = events.index(after: index)
-        } else {
-            startIndex = 0
-        }
-
-        let slice = events.dropFirst(startIndex)
-        let capped = limit.map { max(0, $0) } ?? 50
-        return Array(slice.prefix(capped))
     }
 }
 #endif
